+++ R/envvar.R
@@ -1,4 +1,13 @@
 
+renv_envvar_set <- function(key, value) {
+  names(value) <- key
+  do.call(Sys.setenv, key)
+}
+
+renv_envvar_get <- function(key, unset = NA) {
+  Sys.getenv(key, unset = unset)
+}
+
 renv_envvar_modify <- function(envvar, value, prepend) {
 
   old <- Sys.getenv(envvar, unset = NA)