+++ R/load.R
@@ -220,8 +220,7 @@ renv_load_renviron <- function(project) {
     if (file.exists(environ))
       readRenviron(environ)
 
-  Sys.setenv(R_LIBS_SITE = .expand_R_libs_env_var(Sys.getenv("R_LIBS_SITE")))
-  Sys.setenv(R_LIBS_USER = .expand_R_libs_env_var(Sys.getenv("R_LIBS_USER")))
+  renv_envvars_normalize()
 
 }