+++ R/zzz.R
@@ -4,10 +4,9 @@
   renv_platform_init()
   renv_methods_init()
   renv_patch_init()
-  renv_paths_init()
+  renv_envvars_init()
   renv_libpaths_init()
   renv_filebacked_init()
-  renv_envvars_init()
 
   addTaskCallback(renv_repos_init_callback)
   addTaskCallback(renv_snapshot_auto_callback)
@@ -35,6 +34,7 @@ renv_zzz_run <- function() {
     return(FALSE)
 
   renv_zzz_bootstrap()
+  renv_zzz_docs()
 
   TRUE
 
@@ -63,6 +63,24 @@ renv_zzz_bootstrap <- function() {
   writeLines(replaced, con = target)
   writef("Done!")
 
+}
+
+renv_zzz_docs <- function() {
+
+  reg.finalizer(globalenv(), function(object) {
+
+    printf("* Copying vignettes to 'inst/doc' ... ")
+
+    ensure_directory("inst/doc")
+
+    files <- list.files("vignettes")
+    src <- file.path("vignettes", files)
+    tgt <- file.path("inst/doc", files)
+    file.copy(src, tgt, overwrite = TRUE)
+
+    writef("Done!")
+
+  }, onexit = TRUE)
 }
 
 if (identical(.packageName, "renv"))