+++ tests/testthat/test-paths.R
@@ -38,7 +38,7 @@ test_that("RENV_PATHS_PREFIX is respected", {
 
 test_that("RENV_PATHS_PREFIX is not normalized", {
   renv_scope_envvars(RENV_PATHS_PREFIX = ".")
-  renv_paths_init()
+  renv_envvars_normalize()
   expect_identical(Sys.getenv("RENV_PATHS_PREFIX"), ".")
 })