+++ tests/testthat/test-repair.R
@@ -28,6 +28,7 @@ test_that("we can repair a broken project library", {
 test_that("repair() uses the package version recorded in the lockfile", {
 
   skip_on_cran()
+  renv_scope_options(renv.tests.verbose = FALSE)
   renv_tests_scope("breakfast")
 
   init()